requirements.txt
slackclient
spacy
requests